Fix to include correct Availability.h header
The OS X SDK includes a file named
"Availability.h". This file defines the  macro
__MAC_OS_X_VERSION_MAX_ALLOWED. However, the file
was being imported as "availability.h". On
case-insensitive, case-preserving filesystems,
this name discrepancy is not a problem, but when
compiling on a case-sensitive filesystem, the
compiler fails to find a file named
"availability.h". Include the correct,
case-sensitive filename instead.

Signed-off-by: Link Dupont <link@sub-pop.net>