Improve recovery instruction when ip address not assigned
On managed machines, this happens after every reboot.
`/usr/libexec/bootpd` is already part of the firewall and configured to
accept incoming connections. Removing it from the firewall and adding it
again fixes the issue.

On other machines that do not have this issue, removing and adding
bootpd from the firewall is not needed but also does not cause any
issue.

Fixes: #62
Signed-off-by: Nir Soffer <nsoffer@redhat.com>