feat: Allow specifying vmnet network UUID to disable DHCP (on vmnet.host network only)
This commit introduces a new `--vmnet-network-uuid` command-line option to allow
setting the `vmnet_network_identifier_key` for vmnet.

This property is only applicable to a vmnet_interface in
VMNET_HOST_MODE.

If this property is set, the vmnet_interface is added to an isolated
network with the specified identifier.

No DHCP service is provided on this network.

This is useful for certain applications where the users need an isolated
network and are running their own dhcp to assign IPs in such network.

See issue [#139](https://github.com/lima-vm/socket_vmnet/issues/139)

Signed-off-by: Angelo (pallotron) Failla <pallotron@gmail.com>