test: Wait for iper3 server
Turns out that iperf3 is using the `simple` service type:

    $ sudo systemctl show iperf3 | grep ^Type=
    Type=simple

In this case `systemctl start iperf3` does not wait for the service to
signal that it is ready to handle requests or has finished initializing,
so we may return before the service is ready to serve requests.

We wait now until iperf3 port is in listening state.

Signed-off-by: Nir Soffer <nirsof@gmail.com>